flask/requirements.txt
numpy
Pillow
flask
requests
gunicorn